refactor core into modules
