+++ dustland-core.js
@@ -154,303 +154,6 @@ function genCreatorMap(){
   interiors['creator'] = creatorMap;
 }
 
-// ===== Party / stats =====
-const baseStats = ()=> ({STR:4, AGI:4, INT:4, PER:4, LCK:4, CHA:4});
-
-class Character {
-  constructor(id, name, role){
-    this.id=id; this.name=name; this.role=role;
-    this.lvl=1; this.xp=0;
-    this.stats=baseStats();
-    this.equip={weapon:null, armor:null, trinket:null};
-    this.maxHp=10;
-    this.hp=this.maxHp;
-    this.ap=2;
-    this.map=state.map; this.x=player.x; this.y=player.y;
-    this._bonus={ATK:0, DEF:0, LCK:0};
-  }
-  xpToNext(){ return 10*this.lvl; }
-  awardXP(amt){
-    this.xp += amt;
-    log(`${this.name} gains ${amt} XP.`);
-    if(typeof toast==='function') toast(`${this.name} +${amt} XP`);
-    if(typeof sfxTick==='function') sfxTick();
-    while(this.xp >= this.xpToNext()){
-      this.xp -= this.xpToNext();
-      this.lvl++;
-      this.levelUp();
-    }
-    renderParty(); updateHUD();
-  }
-  levelUp(){
-    const inc = {STR:0,AGI:0,INT:0,PER:0,LCK:0,CHA:0};
-    if(/Gunslinger|Wanderer|Raider/.test(this.role)){ inc.STR++; inc.AGI++; }
-    else if(/Scavenger|Cogwitch|Mechanic/.test(this.role)){ inc.INT++; inc.PER++; }
-    else { inc.CHA++; inc.LCK++; }
-    for(const k in inc){ this.stats[k]+=inc[k]; }
-    this.maxHp += 2;
-    this.hp = Math.min(this.hp + 2, this.maxHp);
-    if(this.lvl%2===0){
-      this.ap += 1;
-      if(typeof hudBadge==='function') hudBadge('AP +1');
-      if(typeof sfxTick==='function') sfxTick();
-    }
-    log(`${this.name} leveled up to ${this.lvl}! (+HP, stats)`);
-  }
-  applyEquipmentStats(){
-    this._bonus = {ATK:0, DEF:0, LCK:0};
-    for(const k of ['weapon','armor','trinket']){
-      const it=this.equip[k];
-      if(it&&it.mods){
-        for(const stat in it.mods){
-          this._bonus[stat]=(this._bonus[stat]||0)+it.mods[stat];
-        }
-      }
-    }
-  }
-}
-
-class Party extends Array {
-  addMember(member){
-    this.push(member);
-    member.applyEquipmentStats();
-    renderParty(); updateHUD();
-    log(member.name+" joins the party.");
-  }
-  leader(){ return this[selectedMember] || this[0]; }
-}
-
-const party = new Party();
-
-function makeMember(id, name, role){ return new Character(id, name, role); }
-function addPartyMember(member){ party.addMember(member); }
-function statLine(s){ return `STR ${s.STR}  AGI ${s.AGI}  INT ${s.INT}  PER ${s.PER}  LCK ${s.LCK}  CHA ${s.CHA}`; }
-function xpToNext(lvl){ return 10*lvl; }
-function awardXP(who, amt){ who.awardXP(amt); }
-function applyEquipmentStats(m){ m.applyEquipmentStats(); }
-function leader(){ return party.leader(); } // handy globally
-
-// ===== Inventory / equipment =====
-const ITEMS = {}; // item definitions by id
-const itemDrops = []; // {map,x,y,id}
-function cloneItem(it){
-  return {
-    ...it,
-    mods: { ...it.mods },
-    tags: [...it.tags],
-    use: it.use ? JSON.parse(JSON.stringify(it.use)) : null,
-    equip: it.equip ? JSON.parse(JSON.stringify(it.equip)) : null
-  };
-}
-function registerItem(item){
-  const norm = normalizeItem(item);
-  if(!norm.id) throw new Error('Item must have id');
-  ITEMS[norm.id] = norm;
-  return norm;
-}
-function getItem(id){
-  const it = ITEMS[id];
-  return it ? cloneItem(it) : null;
-}
-function addToInv(item){
-  let base = null;
-  if(typeof item === 'string'){
-    base = getItem(item);
-  } else if(item && item.id){
-    base = ITEMS[item.id] || registerItem(item);
-    base = cloneItem(base);
-  }
-  if(!base) throw new Error('Unknown item');
-  player.inv.push(base);
-  renderInv();
-  if (window.NanoDialog) {
-    NPCS.filter(n=> n.map === state.map)
-        .forEach(n=> NanoDialog.queueForNPC(n, 'start', 'inventory change'));
-  }
-}
-function removeFromInv(invIndex){
-  player.inv.splice(invIndex,1);
-  renderInv();
-  if (window.NanoDialog) {
-    NPCS.filter(n=> n.map === state.map)
-        .forEach(n=> NanoDialog.queueForNPC(n, 'start', 'inventory change'));
-  }
-}
-let selectedMember = 0;
-
-function equipItem(memberIndex, invIndex){
-  const m=party[memberIndex]; const it=player.inv[invIndex];
-  if(!m||!it||!it.slot){ log('Cannot equip that.'); return; }
-  const slot = it.slot;
-  const prevEq = m.equip[slot];
-  if(prevEq){
-    if(prevEq.cursed){
-      prevEq.cursedKnown = true;
-      renderInv(); renderParty(); updateHUD();
-      log(`${prevEq.name} is cursed and cannot be removed.`);
-      return;
-    }
-    player.inv.push(prevEq);
-  }
-  m.equip[slot]=it;
-  player.inv.splice(invIndex,1);
-  applyEquipmentStats(m);
-  renderInv(); renderParty(); updateHUD();
-  log(`${m.name} equips ${it.name}.`);
-  if(typeof toast==='function') toast(`${m.name} equips ${it.name}`);
-  if(typeof sfxTick==='function') sfxTick();
-  if(it.equip && it.equip.teleport){
-    const t=it.equip.teleport;
-    if(typeof t.x==='number') player.x=t.x;
-    if(typeof t.y==='number') player.y=t.y;
-    if(t.map) setMap(t.map);
-    updateHUD();
-  }
-  if(it.equip && it.equip.msg){
-    log(it.equip.msg);
-  }
-}
-
-function unequipItem(memberIndex, slot){
-  const m=party[memberIndex];
-  if(!m) return;
-  const it=m.equip[slot];
-  if(!it){ log('Nothing to unequip.'); return; }
-  if(it.cursed){
-    it.cursedKnown = true;
-    renderInv(); renderParty(); updateHUD();
-    log(`${it.name} is cursed and won't come off!`);
-    return;
-  }
-  m.equip[slot]=null;
-  player.inv.push(it);
-  applyEquipmentStats(m);
-  renderInv(); renderParty(); updateHUD();
-  log(`${m.name} unequips ${it.name}.`);
-  if(typeof toast==='function') toast(`${m.name} unequips ${it.name}`);
-  if(typeof sfxTick==='function') sfxTick();
-}
-
-function normalizeItem(it){
-  if(!it) return null;
-  return {
-    id: it.id || '',
-    name: it.name || 'Unknown',
-    type: it.type || 'misc',
-    tags: Array.isArray(it.tags) ? it.tags.map(t=>t.toLowerCase()) : [],
-    slot: it.slot || null,
-    mods: it.mods ? { ...it.mods } : {},
-    use: it.use || null,
-    equip: it.equip || null,
-    cursed: !!it.cursed,
-    cursedKnown: !!it.cursedKnown,
-    rarity: it.rarity || 'common',
-    value: Math.max(1, it.value ?? 0),
-    desc: it.desc || '',
-  };
-}
-
-function findItemIndex(idOrTag){
-  const tag = typeof idOrTag === 'string' ? idOrTag.toLowerCase() : '';
-  return player.inv.findIndex(it => it.id === idOrTag || it.tags.map(t=>t.toLowerCase()).includes(tag));
-}
-function hasItem(idOrTag){ return findItemIndex(idOrTag) !== -1; }
-
-function useItem(invIndex){
-  const it = player.inv[invIndex];
-  if(!it || !it.use){
-    log('Cannot use that.');
-    return false;
-  }
-  if(it.use.type==='heal'){
-    const who = (party[selectedMember]||party[0]);
-    if(!who){ log('No party member to heal.'); return false; }
-    const before = who.hp;
-    who.hp = Math.min(who.hp + it.use.amount, who.maxHp);
-    const healed = who.hp - before;
-    log(`${who.name} drinks ${it.name} (+${healed} HP).`);
-    if (typeof toast === 'function') toast(`${who.name} +${healed} HP`);
-    if (typeof sfxTick === 'function') sfxTick();
-    player.inv.splice(invIndex,1);
-    renderInv(); renderParty(); updateHUD();
-    if (window.NanoDialog) {
-      NPCS.filter(n=> n.map === state.map)
-          .forEach(n=> NanoDialog.queueForNPC(n, 'start', 'inventory change'));
-    }
-    return true;
-  }
-  if(typeof it.use.onUse === 'function'){
-    const ok = it.use.onUse({player, party, log, toast});
-    if(ok!==false){
-      player.inv.splice(invIndex,1);
-      renderInv(); renderParty(); updateHUD();
-      if(typeof toast==='function') toast(`Used ${it.name}`);
-      if(typeof sfxTick==='function') sfxTick();
-      if (window.NanoDialog) {
-        NPCS.filter(n=> n.map === state.map)
-            .forEach(n=> NanoDialog.queueForNPC(n, 'start', 'inventory change'));
-      }
-    }
-    return !!ok;
-  }
-  log('Nothing happens...');
-  return false;
-}
-
-// ===== Helpers =====
-function mapIdForState(){ return state.map; }
-function mapWH(map=state.map){
-  if(map==='world') return {W:WORLD_W,H:WORLD_H};
-  if(map==='creator') return {W:creatorMap.w||VIEW_W,H:creatorMap.h||VIEW_H};
-  const I=interiors[map];
-  return {W:(I&&I.w)||VIEW_W,H:(I&&I.h)||VIEW_H};
-}
-function gridFor(map){
-  if(map==='world') return world;
-  if(map==='creator') return creatorMap.grid;
-  return interiors[map] && interiors[map].grid;
-}
-function getTile(map,x,y){
-  const g=gridFor(map); if(!g) return null;
-  if(y<0||x<0||y>=g.length||x>=g[0].length) return null;
-  return g[y][x];
-}
-function setTile(map,x,y,t){
-  const g=gridFor(map); if(!g) return false;
-  if(y<0||x<0||y>=g.length||x>=g[0].length) return false;
-  g[y][x]=t; return true;
-}
-function currentGrid(){ return gridFor(mapIdForState()); }
-function queryTile(x,y,map=mapIdForState()){
-  const tile=getTile(map,x,y);
-  if(tile===null) return {tile:null, walkable:false, entities:[], items:[]};
-  const entities=NPCS.filter(n=> n.map===map && n.x===x && n.y===y);
-  const items=itemDrops.filter(it=> it.map===map && it.x===x && it.y===y);
-  const walkableFlag=!!(walkable[tile] && entities.length===0 && items.length===0);
-  return {tile, walkable:walkableFlag, entities, items};
-}
-// Find nearest free, walkable, unoccupied (and not water on world)
-function findFreeDropTile(map,x,y){
-  const {W,H}=mapWH(map);
-  const dirs=[[1,0],[-1,0],[0,1],[0,-1]];
-  const seen=new Set([x+','+y]);
-  const q=[[x,y,0]];
-  const MAX_RAD=50;
-  while(q.length){
-    const [cx,cy,d]=q.shift();
-    if(d>MAX_RAD) break;
-    if(cx>=0&&cy>=0&&cx<W&&cy<H){
-      const info=queryTile(cx,cy,map);
-      if(info.walkable) return {x:cx,y:cy};
-    }
-    for(const [dx,dy] of dirs){
-      const nx=cx+dx, ny=cy+dy, nd=d+1, k=nx+','+ny;
-      if(nx>=0&&ny>=0&&nx<W&&ny<H&&!seen.has(k) && nd<=MAX_RAD){ seen.add(k); q.push([nx,ny,nd]); }
-    }
-  }
-  return {x,y};
-}
-
 // ===== Quests / NPC =====
 class Quest {
   constructor(id, title, desc, meta={}){
@@ -713,408 +416,6 @@ function placeHut(x,y,b){
   return nb;
 }
 
-// ===== Interaction =====
-function canWalk(x,y){
-  if(state.map==='creator') return false;
-  return queryTile(x,y).walkable;
-}
-function move(dx,dy){
-  if(state.map==='creator') return;
-  const nx=player.x+dx, ny=player.y+dy;
-  if(canWalk(nx,ny)){
-    player.x=nx; player.y=ny;
-    centerCamera(player.x,player.y,state.map); updateHUD();
-  }
-}
-function adjacentNPC(){
-  const dirs=[[1,0],[-1,0],[0,1],[0,-1]];
-  for(const [dx,dy] of dirs){
-    const info=queryTile(player.x+dx,player.y+dy);
-    if(info.entities.length) return info.entities[0];
-  }
-  return null;
-}
-function takeNearestItem(){
-  const dirs=[[0,0],[1,0],[-1,0],[0,1],[0,-1]];
-  for(const [dx,dy] of dirs){
-    const info=queryTile(player.x+dx,player.y+dy);
-    if(info.items.length){
-      const it=info.items[0];
-      const idx=itemDrops.indexOf(it);
-      if(idx>-1) itemDrops.splice(idx,1);
-      const def = ITEMS[it.id];
-      addToInv(it.id);
-      log('Took '+(def?def.name:it.id)+'.'); updateHUD();
-      return true;
-    }
-  }
-  return false;
-}
-function interactAt(x,y){
-  if(state.map==='creator') return false;
-  const dist=Math.abs(player.x-x)+Math.abs(player.y-y);
-  if(dist>1) return false;
-  const info=queryTile(x,y);
-  if(info.entities.length){ openDialog(info.entities[0]); return true; }
-  if(info.items.length){
-    const it=info.items[0];
-    const idx=itemDrops.indexOf(it);
-    if(idx>-1) itemDrops.splice(idx,1);
-    const def = ITEMS[it.id];
-    addToInv(it.id); log('Took '+(def?def.name:it.id)+'.'); updateHUD();
-    return true;
-  }
-  if(x===player.x && y===player.y && info.tile===TILE.DOOR){
-    if(state.map==='world'){
-      const b=buildings.find(b=> b.doorX===x && b.doorY===y);
-      if(!b){ log('No entrance here.'); return true; }
-      if(b.boarded){ log('The doorway is boarded up from the outside.'); return true; }
-      const I=interiors[b.interiorId];
-      if(I){ player.x=I.entryX; player.y=I.entryY; }
-      setMap(b.interiorId,'Interior');
-      log('You step inside.'); updateHUD(); return true;
-    }
-    if(state.map!=='world'){
-      const p=portals.find(p=> p.map===state.map && p.x===x && p.y===y);
-      if(p){
-        const target=p.toMap;
-        const I=interiors[target];
-        player.x = (typeof p.toX==='number') ? p.toX : (I?I.entryX:0);
-        player.y = (typeof p.toY==='number') ? p.toY : (I?I.entryY:0);
-        setMap(target);
-        log(p.desc || 'You move through the doorway.');
-        updateHUD(); return true;
-      }
-      const b=buildings.find(b=> b.interiorId===state.map);
-      if(b){ player.x=b.doorX; player.y=b.doorY-1; setMap('world'); log('You step back outside.'); updateHUD(); return true; }
-    }
-  }
-  return false;
-}
-function interact(){
-  if(Date.now()-lastInteract < 200) return false;
-  lastInteract = Date.now();
-  const dirs=[[0,0],[1,0],[-1,0],[0,1],[0,-1]];
-  for(const [dx,dy] of dirs){
-    if(interactAt(player.x+dx,player.y+dy)) return true;
-  }
-  log('Nothing interesting.');
-  return false;
-}
-
-// Grouped systems expose focused gameplay concerns
-const movementSystem = { canWalk, move };
-const collisionSystem = { queryTile, canWalk };
-const interactionSystem = { adjacentNPC, takeNearestItem, interact, interactAt };
-Object.assign(window, { movementSystem, collisionSystem, interactionSystem, queryTile, interactAt });
-
-// ===== Dialog =====
-const overlay=document.getElementById('overlay');
-const choicesEl=document.getElementById('choices');
-const textEl=document.getElementById('dialogText');
-const nameEl=document.getElementById('npcName');
-const titleEl=document.getElementById('npcTitle');
-const portEl=document.getElementById('port');
-let currentNPC=null;
-const dialogState={ tree:null, node:null };
-
-function normalizeDialogTree(tree){
-  const out={};
-  for(const id in tree){
-    const n=tree[id];
-    const next=(n.next||n.choices||[]).map(c=>{
-      if(typeof c==='string') return {id:c,label=c};
-      const {to,id:cid,label,text,checks=[],effects=[],...rest}=c;
-      return {id:to||cid,label:label||text||'(Continue)',checks,effects,...rest};
-    });
-    out[id]={text:n.text||'',checks:n.checks||[],effects:n.effects||[],next};
-  }
-  return out;
-}
-
-function runEffects(effects){
-  for(const fn of effects||[]){ if(typeof fn==='function') fn({player,party,state}); }
-}
-
-/**
- * Resolve a stat check against a DC and run effects.
- * @param {Check} check
- * @param {Character} [actor=leader()]
- * @param {Function} [rng=Math.random]
- * @returns {{success:boolean, roll:number, dc:number, stat:string}}
- */
-function resolveCheck(check, actor=leader(), rng=Math.random){
-  const roll = Dice.skill(actor, check.stat, 0, ROLL_SIDES, rng);
-  const dc = check.dc || 0;
-  const success = roll >= dc;
-  log?.(`Check ${check.stat} rolled ${roll} vs DC ${dc}: ${success?'success':'fail'}`);
-  runEffects(success ? check.onSuccess : check.onFail);
-  return { success, roll, dc, stat: check.stat };
-}
-
-function applyReward(reward){
-  if(!reward) return;
-  if(typeof reward==='string' && /^xp\s*\d+/i.test(reward)){
-    const amt=parseInt(reward.replace(/[^0-9]/g,''),10)||0;
-    awardXP(leader(), amt);
-    if(typeof toast==='function') toast(`+${amt} XP`);
-  } else {
-    addToInv(reward);
-    const id = typeof reward === 'string' ? reward : reward.id;
-    const def = ITEMS[id] || (typeof reward === 'object' ? reward : null);
-    if(typeof toast==='function') toast(`Received ${def?.name || id}`);
-  }
-}
-
-function processQuestFlag(c){
-  if(!currentNPC?.quest || !c?.q) return;
-  if(c.q==='accept') defaultQuestProcessor(currentNPC,'accept');
-  if(c.q==='turnin') defaultQuestProcessor(currentNPC,'do_turnin');
-}
-
-function joinParty(join){
-  if(!join) return;
-  const m=makeMember(join.id, join.name, join.role);
-  addPartyMember(m);
-  removeNPC(currentNPC);
-}
-
-function handleGoto(g){
-  if(!g) return;
-  if(g.map==='world'){
-    startWorld();
-    if(typeof g.x==='number') player.x=g.x;
-    if(typeof g.y==='number') player.y=g.y;
-    setMap('world');
-  }else{
-    if(typeof g.x==='number') player.x=g.x;
-    if(typeof g.y==='number') player.y=g.y;
-    if(g.map) setMap(g.map); else if(typeof centerCamera==='function') centerCamera(player.x,player.y,state.map);
-  }
-  updateHUD?.();
-}
-
-function advanceDialog(stateObj, choiceIdx){
-  const node=stateObj.tree[stateObj.node];
-  const choice=node.next[choiceIdx];
-  if(!choice){ stateObj.node=null; return {next:null, text:null, close:true}; }
-
-  // pre-check effects (e.g., condition gates)
-  runEffects(choice.checks);
-
-  const res={next:null, text:null, close:false};
-  const finalize=(text)=>{ res.text=text||null; res.close=true; stateObj.node=null; return res; };
-
-  // Required item/slot without consumption
-  if(choice.reqItem || choice.reqSlot){
-    const idx = choice.reqItem
-      ? findItemIndex(choice.reqItem)
-      : player.inv.findIndex(it => it.slot === choice.reqSlot);
-    if(idx === -1){
-      return finalize(choice.failure || 'You lack the required item.');
-    }
-    applyReward(choice.reward);
-    joinParty(choice.join);
-    processQuestFlag(choice);
-    runEffects(choice.effects);
-    if(choice.goto){
-      handleGoto(choice.goto);
-      res.close=true;
-      stateObj.node=null;
-      return res;
-    }
-    return finalize(choice.success || '');
-  }
-
-  // Item/slot costs
-  if(choice.costItem || choice.costSlot){
-    const idx = choice.costItem ? findItemIndex(choice.costItem)
-                                : player.inv.findIndex(it=> it.slot===choice.costSlot);
-    if(idx === -1){
-      return finalize(choice.failure || 'You lack the required item.');
-    }
-    removeFromInv(idx);
-    applyReward(choice.reward);
-    joinParty(choice.join);
-    processQuestFlag(choice);
-    runEffects(choice.effects);
-    if(choice.goto){
-      handleGoto(choice.goto);
-      res.close=true;
-      stateObj.node=null;
-      return res;
-    }
-    return finalize(choice.success || '');
-  }
-
-  // Stat roll gates
-  if(choice.check){
-    const { success, roll, dc } = resolveCheck(choice.check, leader());
-    if(success){
-      applyReward(choice.reward);
-      joinParty(choice.join);
-      processQuestFlag(choice);
-    }
-    runEffects(choice.effects);
-    const msg = (success?choice.success:choice.failure)||'';
-    return finalize(msg + ` (Roll ${roll} vs DC ${dc}).`);
-  }
-
-  // Direct response
-  if(choice.response){
-    processQuestFlag(choice);
-    runEffects(choice.effects);
-    return finalize(choice.response);
-  }
-
-  // Reward/join without roll
-  if(choice.reward || choice.join){
-    applyReward(choice.reward);
-    joinParty(choice.join);
-    processQuestFlag(choice);
-    runEffects(choice.effects);
-    return finalize(choice.success || '');
-  }
-
-  // Map moves
-  if(choice.goto){
-    handleGoto(choice.goto);
-    processQuestFlag(choice);
-    runEffects(choice.effects);
-    res.close=true;
-    stateObj.node=null;
-    return res;
-  }
-
-  // Default: move to another node
-  processQuestFlag(choice);
-  runEffects(choice.effects);
-  stateObj.node=choice.id||null;
-  res.next=stateObj.node;
-  if(!res.next) res.close=true;
-  return res;
-}
-
-// Portrait selection from a 2x2 sheet or fallback color
-function setPortrait(portEl, npc){
-  if(!npc.portraitSheet){
-    portEl.style.backgroundImage = '';
-    portEl.style.background = npc.color || '#274227';
-    return;
-  }
-  const frame = Math.floor(Math.random() * 4); // 0..3
-  const col = frame % 2;         // 0 or 1
-  const row = Math.floor(frame/2); // 0 or 1
-  const posX = col === 0 ? '0%'   : '100%';
-  const posY = row === 0 ? '0%'   : '100%';
-  portEl.style.background = 'transparent';
-  portEl.style.backgroundImage = `url(${npc.portraitSheet})`;
-  portEl.style.backgroundSize = '200% 200%';
-  portEl.style.backgroundPosition = `${posX} ${posY}`;
-}
-
-function openDialog(npc, node='start'){
-  currentNPC=npc;
-  dialogState.tree=normalizeDialogTree(npc.tree||{});
-  dialogState.node=node;
-  nameEl.textContent=npc.name;
-  titleEl.textContent=npc.title;
-
-  setPortrait(portEl, npc);
-
-  // Optional subtitle/desc adornment
-  const desc = npc.desc;
-  if(desc){
-    const small=document.createElement('div');
-    small.className='small npcdesc';
-    small.textContent=desc;
-    const hdr=titleEl.parentElement;
-    [...hdr.querySelectorAll('.small.npcdesc')].forEach(n=>n.remove());
-    hdr.appendChild(small);
-  }
-
-  renderDialog();
-  overlay.classList.add('shown');
-  setGameState(GAME_STATE.DIALOG);
-}
-
-function closeDialog(){
-  overlay.classList.remove('shown');
-  currentNPC=null;
-  dialogState.tree=null;
-  dialogState.node=null;
-  choicesEl.innerHTML='';
-  const back= state.map==='world'?GAME_STATE.WORLD:GAME_STATE.INTERIOR;
-  setGameState(back);
-}
-
-function renderDialog(){
-  if(!dialogState.tree) return;
-  const node=dialogState.tree[dialogState.node];
-  if(!node){ closeDialog(); return; }
-
-  runEffects(node.checks);
-  runEffects(node.effects);
-
-  textEl.textContent=node.text;
-  choicesEl.innerHTML='';
-
-  if(!node.next || node.next.length===0){
-    const cont=document.createElement('div');
-    cont.className='choice';
-    cont.textContent='(Continue)';
-    cont.onclick=()=> closeDialog();
-    choicesEl.appendChild(cont);
-    return;
-  }
-
-  // Start with all choices, then apply filters
-  let choices=node.next.map((opt,idx)=>({opt,idx}));
-
-  // Quest choice visibility (accept/turnin)
-  if(currentNPC?.quest){
-    const meta=currentNPC.quest;
-    choices=choices.filter(({opt})=>{
-      if(opt.q==='accept' && meta.status!=='available') return false;
-      if(opt.q==='turnin' && (meta.status==='completed' || (meta.item && !hasItem(meta.item)))) return false;
-      return true;
-    });
-  }
-
-  // Once-only choices
-  choices=choices.filter(({opt})=>{
-    if(!opt.once) return true;
-    const key=`${currentNPC.id}::${dialogState.node}::${opt.label}`;
-    return !usedOnceChoices.has(key);
-  });
-
-  // Render choices
-  choices.forEach(({opt,idx})=>{
-    const div=document.createElement('div');
-    div.className='choice';
-    div.textContent=opt.label||'(Continue)';
-    div.onclick=()=>{
-      const key=`${currentNPC.id}::${dialogState.node}::${opt.label}`;
-      if(opt.once) usedOnceChoices.add(key);
-
-      const result=advanceDialog(dialogState,idx);
-      if(result && result.text!==null){
-        textEl.textContent=result.text;
-        choicesEl.innerHTML='';
-        const cont=document.createElement('div');
-        cont.className='choice';
-        cont.textContent='(Continue)';
-        cont.onclick=()=>{ if(result.close) closeDialog(); else { dialogState.node=result.next; renderDialog(); } };
-        choicesEl.appendChild(cont);
-      } else {
-        if(result && result.close) closeDialog();
-        else renderDialog();
-      }
-    };
-    choicesEl.appendChild(div);
-  });
-}
-
 // ===== Save/Load & Start =====
 function save(){
   const npcData = NPCS.map(({id,map,x,y,quest})=>({id,map,x,y,quest:quest?{id:quest.id,status:quest.status}:null}));
@@ -1297,41 +598,17 @@ function startWorld(){
 
 // Content pack moved to modules/dustland.module.js
 
-Object.assign(window, {Dice, Character, Party, Quest, NPC, questLog, quickCombat, removeNPC, makeNPC, resolveCheck});
 
-// Export a few helpers for Node-based tests without affecting the browser build
+// Export a few helpers for Node-based tests without affecting the browser build
+
+const coreExports = { ROLL_SIDES, clamp, createRNG, Dice, quickCombat, TILE, walkable, mapLabels, mapLabel, setMap, isWalkable, VIEW_W, VIEW_H, TS, WORLD_W, WORLD_H, world, interiors, buildings, portals, state, player, GAME_STATE, setGameState, doorPulseUntil, lastInteract, creatorMap, genCreatorMap, Quest, NPC, questLog, NPCS, addQuest, completeQuest, defaultQuestProcessor, removeNPC, makeNPC, applyModule, genWorld, isWater, findNearestLand, makeInteriorRoom, placeHut, startGame, startWorld };
+
+Object.assign(globalThis, coreExports);
+
 if (typeof module !== 'undefined' && module.exports) {
-  module.exports = {
-    addToInv,
-    advanceDialog,
-    applyModule,
-    canWalk,
-    Character,
-    clamp,
-    createRNG,
-    Dice,
-    equipItem,
-    findFreeDropTile,
-    GAME_STATE,
-    getGameState: () => gameState,
-    itemDrops,
-    move,
-    normalizeItem,
-    interactAt,
-    registerItem,
-    queryTile,
-    NPCS,
-    openDialog,
-    party,
-    portals,
-    player,
-    ITEMS,
-    mapLabels,
-    takeNearestItem,
-    resolveCheck,
-    setGameState,
-    setLeader: (idx)=>{ selectedMember = idx; },
-    state,
-    unequipItem
-  };
+  const party = require('./core/party.js');
+  const inventory = require('./core/inventory.js');
+  const movement = require('./core/movement.js');
+  const dialog = require('./core/dialog.js');
+  module.exports = { ...coreExports, ...party, ...inventory, ...movement, ...dialog, getGameState: () => gameState };
 }
