+++ dustland.html
@@ -100,6 +100,10 @@
   </div>
 
   <script src="dustland-core.js"></script>
+  <script src="core/party.js"></script>
+  <script src="core/inventory.js"></script>
+  <script src="core/movement.js"></script>
+  <script src="core/dialog.js"></script>
   <script>
     window._realOpenCreator = window.openCreator;
     window._realShowStart = window.showStart;
